store money as an integer

+++ app/schemas/record_schema.py
@@ -8,7 +8,7 @@ class RecordSchema(Schema):
     id = fields.Int(dump_only=True)
     user_id = fields.Int(required=True, validate=validate.Range(min=1))
     category_id = fields.Int(required=True, validate=validate.Range(min=1))
-    sum = fields.Float(required=True, validate=validate.Range(min=min_sum))
+    sum = fields.Int(required=True, validate=validate.Range(min=min_sum))
     date_time = fields.DateTime(dump_only=True)
     # user = fields.Nested('UserSchema', dump_only=True)
     # category = fields.Nested('CategorySchema', dump_only=True)
@@ -18,5 +18,5 @@ class RecordCreateSchema(Schema):
     """Schema for creating a new record"""
     user_id = fields.Int(required=True, validate=validate.Range(min=1))
     category_id = fields.Int(required=True, validate=validate.Range(min=1))
-    sum = fields.Float(required=True, validate=validate.Range(min=min_sum))
+    sum = fields.Int(required=True, validate=validate.Range(min=min_sum))
     date_time = fields.DateTime(load_default=datetime.utcnow, allow_none=True)
